+++ src/components/user-action/DateAndTimePicker.tsx
@@ -1,12 +1,10 @@
 import type { ReactNode } from 'react'
 import clsx from 'clsx'
 import { addDuration, subtractDuration } from '@/src/utils/date'
-import { Button } from './Button'
 import type { TimePickerProps } from '../date/TimePicker'
 import { TimePicker } from '../date/TimePicker'
 import type { DatePickerProps } from '../date/DatePicker'
 import { DatePicker } from '../date/DatePicker'
-import { useHightideTranslation } from '@/src/i18n/useHightideTranslation'
 
 export type DateTimePickerMode = 'date' | 'time' | 'dateTime'
 
@@ -16,8 +14,6 @@ export type DateTimePickerProps = {
   start?: Date,
   end?: Date,
   onChange?: (date: Date) => void,
-  onFinish?: (date: Date) => void,
-  onRemove?: () => void,
   datePickerProps?: Omit<DatePickerProps, 'onChange' | 'value' | 'start' | 'end'>,
   timePickerProps?: Omit<TimePickerProps, 'onChange' | 'time' | 'maxHeight'>,
 }
@@ -30,14 +26,10 @@ export const DateTimePicker = ({
                                  start = subtractDuration(new Date(), { years: 50 }),
                                  end = addDuration(new Date(), { years: 50 }),
                                  mode = 'dateTime',
-                                 onFinish,
                                  onChange,
-                                 onRemove,
                                  timePickerProps,
                                  datePickerProps,
                                }: DateTimePickerProps) => {
-  const translation = useHightideTranslation()
-
   const useDate = mode === 'dateTime' || mode === 'date'
   const useTime = mode === 'dateTime' || mode === 'time'
 
@@ -48,7 +40,7 @@ export const DateTimePicker = ({
     dateDisplay = (
       <DatePicker
         {...datePickerProps}
-        className="min-w-80 min-h-71 max-h-71"
+        className="min-w-80"
         yearMonthPickerProps={{ className: 'h-full grow' }}
         value={value}
         start={start}
@@ -61,7 +53,7 @@ export const DateTimePicker = ({
     timeDisplay = (
       <TimePicker
         {...timePickerProps}
-        className={clsx('min-h-71 max-h-71', { 'justify-between w-full': mode === 'time' })}
+        className={clsx({ 'justify-between': mode === 'time' })}
         time={value}
         onChange={onChange}
       />
@@ -69,22 +61,9 @@ export const DateTimePicker = ({
   }
 
   return (
-    <div className="flex-col-2 w-fit">
-      <div className="flex-row-4">
-        {dateDisplay}
-        {timeDisplay}
-      </div>
-      <div className="flex-row-2 justify-end">
-        <div className="flex-row-2 mt-1">
-          <Button size="medium" color="negative" onClick={onRemove}>{translation('clear')}</Button>
-          <Button
-            size="medium"
-            onClick={() => onFinish?.(value)}
-          >
-            {translation('change')}
-          </Button>
-        </div>
-      </div>
+    <div className="flex-row-2 min-h-71 max-h-71">
+      {dateDisplay}
+      {timeDisplay}
     </div>
   )
 }