+++ src/components/user-action/Tooltip.tsx
@@ -1,7 +1,10 @@
 import type { CSSProperties, PropsWithChildren, ReactNode } from 'react'
-import { useHoverState } from '../../hooks/useHoverState'
+import { useRef, useState } from 'react'
 import { clsx } from 'clsx'
 import { useZIndexRegister } from '@/src/hooks/useZIndexRegister'
+import { Visibility } from '@/src/components/layout/Visibility'
+import { useFloatingElement } from '@/src/hooks/useFloatingElement'
+import { createPortal } from 'react-dom'
 
 type Position = 'top' | 'bottom' | 'left' | 'right'
 
@@ -10,9 +13,15 @@ export type TooltipProps = PropsWithChildren<{
   /**
    * Number of milliseconds until the tooltip appears
    *
-   * defaults to 1000ms
+   * defaults to 400ms
    */
-  animationDelay?: number,
+  appearDelay?: number,
+  /**
+   * Number of milliseconds until the tooltip disappears
+   *
+   * defaults to 50ms
+   */
+  disappearDelay?: number,
   /**
    * Class names of additional styling properties for the tooltip
    */
@@ -22,7 +31,13 @@ export type TooltipProps = PropsWithChildren<{
    */
   containerClassName?: string,
   position?: Position,
+  disabled?: boolean,
 }>
+
+type TooltipState = {
+  isShown: boolean,
+  timer: NodeJS.Timeout | null,
+}
 
 /**
  * A Component for showing a tooltip when hovering over Content
@@ -37,59 +52,117 @@ export type TooltipProps = PropsWithChildren<{
 export const Tooltip = ({
                           tooltip,
                           children,
-                          animationDelay = 650,
+                          appearDelay = 400,
+                          disappearDelay = 50,
                           tooltipClassName = '',
                           containerClassName = '',
                           position = 'bottom',
+                          disabled = false,
                         }: TooltipProps) => {
-  const { isHovered, handlers } = useHoverState()
+  const [state, setState] = useState<TooltipState>({
+    isShown: false,
+    timer: null,
+  })
+  const { isShown } = state
+  const anchorRef = useRef<HTMLDivElement>(null)
+  const containerRef = useRef<HTMLDivElement>(null)
+  const triangleRef = useRef<HTMLDivElement>(null)
 
-  const positionClasses = {
-    top: `bottom-full left-1/2 -translate-x-1/2 mb-[6px]`,
-    bottom: `top-full left-1/2 -translate-x-1/2 mt-[6px]`,
-    left: `right-full top-1/2 -translate-y-1/2 mr-[6px]`,
-    right: `left-full top-1/2 -translate-y-1/2 ml-[6px]`
-  }
-
-  const triangleSize = 6
+  const triangleSize = 0.375
   const triangleClasses = {
-    top: `top-full left-1/2 -translate-x-1/2 border-t-tooltip-background border-l-transparent border-r-transparent`,
-    bottom: `bottom-full left-1/2 -translate-x-1/2 border-b-tooltip-background border-l-transparent border-r-transparent`,
-    left: `left-full top-1/2 -translate-y-1/2 border-l-tooltip-background border-t-transparent border-b-transparent`,
-    right: `right-full top-1/2 -translate-y-1/2 border-r-tooltip-background border-t-transparent border-b-transparent`
+    top: `border-t-tooltip-background border-l-transparent border-r-transparent`,
+    bottom: `border-b-tooltip-background border-l-transparent border-r-transparent`,
+    left: `border-l-tooltip-background border-t-transparent border-b-transparent`,
+    right: `border-r-tooltip-background border-t-transparent border-b-transparent`
   }
 
   const triangleStyle: Record<Position, CSSProperties> = {
-    top: { borderWidth: `${triangleSize}px ${triangleSize}px 0 ${triangleSize}px` },
-    bottom: { borderWidth: `0 ${triangleSize}px ${triangleSize}px ${triangleSize}px` },
-    left: { borderWidth: `${triangleSize}px 0 ${triangleSize}px ${triangleSize}px` },
-    right: { borderWidth: `${triangleSize}px ${triangleSize}px ${triangleSize}px 0` }
+    top: { borderWidth: `${triangleSize}rem ${triangleSize}rem 0 ${triangleSize}rem`, transform: `translate(0,${triangleSize}rem)` },
+    bottom: { borderWidth: `0 ${triangleSize}rem ${triangleSize}rem ${triangleSize}rem`, transform: `translate(0,-${triangleSize}rem)` },
+    left: { borderWidth: `${triangleSize}rem 0 ${triangleSize}rem ${triangleSize}rem`, transform: `translate(${triangleSize}rem,0)` },
+    right: { borderWidth: `${triangleSize}rem ${triangleSize}rem ${triangleSize}rem 0`, transform: `translate(-${triangleSize}rem,0)` }
   }
 
-  const zIndex = useZIndexRegister(isHovered)
+  const isActive = !disabled && isShown
+
+  const css = useFloatingElement({
+    active: isActive,
+    anchorRef: anchorRef,
+    containerRef,
+    horizontalAlignment: position === 'left' ? 'beforeStart' : position === 'right' ? 'afterEnd' : 'center',
+    verticalAlignment: position === 'top' ? 'beforeStart' : position === 'bottom' ? 'afterEnd' : 'center',
+  })
+
+  const cssTriangle = useFloatingElement({
+    active: isActive,
+    anchorRef: anchorRef,
+    containerRef: triangleRef,
+    horizontalAlignment: position === 'left' ? 'beforeStart' : position === 'right' ? 'afterEnd' : 'center',
+    verticalAlignment: position === 'top' ? 'beforeStart' : position === 'bottom' ? 'afterEnd' : 'center',
+  })
+
+  const zIndex = useZIndexRegister(isActive)
+  const zIndexTriangle = useZIndexRegister(isActive)
 
   return (
     <div
+      ref={anchorRef}
       className={clsx('relative inline-block', containerClassName)}
-      {...handlers}
+      onMouseEnter={() => setState(prevState => {
+        clearTimeout(prevState.timer)
+        return {
+          ...prevState,
+          timer: setTimeout(() => {
+            setState(prevState => {
+              clearTimeout(prevState.timer)
+              return { isShown: true, timer: null }
+            })
+          }, appearDelay)
+        }
+      })}
+      onMouseLeave={() => setState(prevState => {
+        clearTimeout(prevState.timer)
+        return {
+          ...prevState,
+          timer: setTimeout(() => {
+            setState(prevState => {
+              clearTimeout(prevState.timer)
+              return { isShown: false, timer: null }
+            })
+          }, disappearDelay)
+        }
+      })}
     >
       {children}
-      {isHovered && (
-        <div
-          className={clsx(
-            `opacity-0 absolute text-xs font-semibold text-tooltip-text px-2 py-1 rounded whitespace-nowrap
-           animate-tooltip-fade-in shadow-around-md bg-tooltip-background`,
-            positionClasses[position], tooltipClassName
-          )}
-          style={{ zIndex, animationDelay: animationDelay + 'ms' }}
-        >
-          {tooltip}
+      <Visibility isVisible={isActive}>
+        {createPortal(
           <div
-            className={clsx(`absolute w-0 h-0`, triangleClasses[position])}
-            style={{ ...triangleStyle[position], zIndex: zIndex + 1 }}
+            ref={containerRef}
+            className={clsx(
+              `opacity-0 absolute text-xs font-semibold text-tooltip-text px-2 py-1 rounded
+           animate-tooltip-fade-in shadow-around-md bg-tooltip-background`,
+              tooltipClassName
+            )}
+            style={{ ...css, zIndex, animationDelay: appearDelay + 'ms' }}
+          >
+            {tooltip}
+          </div>
+          , document.body
+        )}
+        {createPortal(
+          <div
+            ref={triangleRef}
+            className={clsx(`absolute w-0 h-0 opacity-0 animate-tooltip-fade-in`, triangleClasses[position])}
+            style={{
+              ...cssTriangle,
+              ...triangleStyle[position],
+              zIndex: zIndexTriangle,
+              animationDelay: appearDelay + 'ms'
+          }}
           />
-        </div>
-      )}
+          , document.body
+        )}
+      </Visibility>
     </div>
   )
 }